decode 路径自洽性验证: 修复 self-check 测试设计错误 + SSM state 改 FP32
- 新测试 model_forward_dump_decode: prefill + greedy decode logits dump,
  Q4T_DECODE_SELFCHK 自洽检查 (prefill(T+1) 末行 vs prefill(T)+decode(1))
- 根因: self-check 误用 decode step 0 的输出 token 构造 prefill5,
  与 decode 实际输入 (prefill argmax) 不同 → 两条路径处理不同 token,
  初测 cos 0.41/0.53 为测试设计错误, 非引擎 bug
- 修复: 用 decode 输入 token 构造 p5; 修复后 1 层 bit-identical
  (全部中间量 max_abs=0), 3 层 cos=0.9985 (top-2 差 0.031, BF16 噪声),
  4 层 cos=0.9988 且 argmax 一致
- SSM 持久 state BF16→FP32 (匹配 transformers 参考全程 FP32 递归;
  使 SSM state 两路径 bit-identical; 36 层 36→72 MB 可忽略)
- 调试设施: Q4T_STATE_DUMP (layer0 ssm/conv/trunk dump) +
  Q4T_LIN_DUMP (linear attention 中间量 x/qkv_raw/qkv/y_ssm dump)
- 修复 decoder_layer 测试 ResetLinearState 清零大小 (×2→×4)
- 54 项测试全绿, 零警告

+++ include/q4t/model/decoder_layer.h
@@ -68,7 +68,7 @@ struct DecoderLayer {
   HyperConnectionWeights mlp_hc;
 
   // Per-layer persistent caches (allocated in Load, freed in Free).
-  uint16_t* ssm_state = nullptr;  // linear: [nv, kd, vd]
+  float* ssm_state = nullptr;  // linear: [nv, kd, vd] (FP32, matches reference)
   uint16_t* conv_state = nullptr;  // linear: [in_qkv, conv_k-1]
   uint16_t* kv_cache = nullptr;  // full: paged [n_pages, kKvPageSize, nkv, 2, hd]
   int* page_table = nullptr;  // full: [max_len] logical pos -> physical page
